Title: 합과 차의 나눗셈, Time: 152ms, Memory: 10MB, Status: Passed - CodeTree

+++ 241004/합과 차의 나눗셈/divide-of-sum-and-sub.java
@@ -7,6 +7,6 @@ public class Main {
 
         int a = sc.nextInt(), b = sc.nextInt();
 
-        System.out.printf("%.2f", (a+b) / (a-b));
+        System.out.printf("%.2f", (double)(a+b) / (a-b));
     }
 }
